+++ decentraland/screens/helpers.js
@@ -1,8 +1,5 @@
 import { Action, ContractBasedAccount, TasitContracts } from "tasit-sdk";
 
-import AssetTypes from "@constants/AssetTypes";
-const { ESTATE, PARCEL } = AssetTypes;
-
 import ProviderFactory from "tasit-action/dist/ProviderFactory";
 import { createFromPrivateKey } from "tasit-account/dist/testHelpers/helpers";
 
@@ -64,69 +61,8 @@ export const addressesAreEqual = (address1, address2) => {
   return address1.toUpperCase() === address2.toUpperCase();
 };
 
-export const prepareEstateForSale = async (estateContract, estateForSale) => {
-  const { id, assetId, seller, priceInWei, expiresAt } = estateForSale;
-
-  const estateId = Number(assetId);
-
-  // Note: Conversion to USD will be implemented on v0.2.0
-  const manaPerUsd = 30;
-  const priceMana = Number(`${priceInWei}`) / 1e18;
-  const priceUSD = Number(priceMana / manaPerUsd).toFixed(2);
-  const name = await estateContract.getMetadata(assetId);
-  const imgUrl = `https://api.decentraland.org/v1/estates/${estateId}/map.png`;
-
-  return {
-    id,
-    priceMana,
-    priceUSD,
-    seller,
-    expiresAt,
-    type: ESTATE,
-    asset: {
-      id: estateId,
-      name,
-      img: imgUrl,
-    },
-  };
-};
-
-export const prepareParcelForSale = async (landContract, parcelForSale) => {
-  const {
-    id,
-    assetId: parcelId,
-    seller,
-    priceInWei,
-    expiresAt,
-  } = parcelForSale;
-
-  // Note: Conversion to USD will be implemented on v0.2.0
-  const manaPerUsd = 30;
-  const priceMana = Number(`${priceInWei}`) / 1e18;
-  const priceUSD = Number(priceMana / manaPerUsd).toFixed(2);
-  const namePromise = landContract.tokenMetadata(parcelId);
-  const coordsPromise = landContract.decodeTokenId(parcelId);
-  const [name, coords] = await Promise.all([namePromise, coordsPromise]);
-  const [x, y] = coords;
-  const imgUrl = `https://api.decentraland.org/v1/parcels/${x}/${y}/map.png`;
-
-  return {
-    id,
-    priceMana,
-    priceUSD,
-    seller,
-    expiresAt,
-    type: PARCEL,
-    asset: {
-      id: parcelId,
-      name,
-      img: imgUrl,
-    },
-  };
-};
-
 export const fundAccount = async accountAddress => {
-  const SMALL_AMOUNT = `${1e17}`; // 0.1
+  const SMALL_AMOUNT = `${5e16}`; // 0.1
   const TEN = `${10e18}`;
 
   const gnosisSafeOwnerPrivKey =
@@ -160,8 +96,6 @@ export const showInfo = msg => console.warn(`INFO: ${msg}`);
 export const showSuccess = msg => console.warn(`SUCCESS: ${msg}`);
 
 export default {
-  prepareParcelForSale,
-  prepareEstateForSale,
   approveManaSpending,
   addressesAreEqual,
   showFatalError,
